+++ CMakeLists.txt
@@ -151,6 +151,8 @@ set(LIB_SOURCE_FILES
         src/dvr.cpp
         src/mpp_encoder.h
         src/mpp_encoder.cpp
+        src/encoder_pacer.h
+        src/encoder_pacer.cpp
         src/mavlink.h
         src/mavlink.c
         src/wfbcli.hpp
